@uppy/dashboard: refactor to ESM
