+++ packages/@uppy/dashboard/src/components/AddFilesPanel.jsx
@@ -1,6 +1,6 @@
-const { h } = require('preact')
-const classNames = require('classnames')
-const AddFiles = require('./AddFiles')
+import { h } from 'preact'
+import classNames from 'classnames'
+import AddFiles from './AddFiles.jsx'
 
 const AddFilesPanel = (props) => {
   return (
@@ -26,4 +26,4 @@ const AddFilesPanel = (props) => {
   )
 }
 
-module.exports = AddFilesPanel
+export default AddFilesPanel
