+++ packages/@uppy/dashboard/src/utils/copyToClipboard.test.js
@@ -1,4 +1,5 @@
-const copyToClipboard = require('./copyToClipboard')
+import { describe, xit, expect } from '@jest/globals'
+import copyToClipboard from './copyToClipboard.js'
 
 describe('copyToClipboard', () => {
   xit('should copy the specified text to the clipboard', () => {
