+++ packages/@uppy/dashboard/src/utils/getActiveOverlayEl.js
@@ -1,7 +1,7 @@
 /**
  * @returns {HTMLElement} - either dashboard element, or the overlay that's most on top
  */
-module.exports = function getActiveOverlayEl (dashboardEl, activeOverlayType) {
+export default function getActiveOverlayEl (dashboardEl, activeOverlayType) {
   if (activeOverlayType) {
     const overlayEl = dashboardEl.querySelector(`[data-uppy-paneltype="${activeOverlayType}"]`)
     // if an overlay is already mounted
